backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# ML/AI Dependencies
torch==2.1.1
transformers==4.35.2
peft==0.6.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
datasets==2.14.6
accelerate==0.24.1

# Audio processing
openai-whisper==20231117
librosa==0.10.1

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP requests for external APIs
httpx==0.25.2
requests==2.31.0

# Development tools
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Logging and monitoring
loguru==0.7.2

# Additional utilities
numpy==1.24.3
scipy==1.11.4
pandas==2.1.3
python-dateutil==2.8.2